FixedPoint improvements (#5)
* fix some lint errors

uint256 -> uint

* make sqrt less lossy

uint256 -> uint and less explicit casting

* add some uint112 boundary tests to sqrt

* add muli

tweak require messages

optimize uint container sizes

* x -> y

* fix muli

add muli tests

* add comment about lossiness

* add boundary tests for mul/muli

* remove div

revert on reciprocal overflow

add more comment notes

* address review comments

* prettier

+++ contracts/test/TransferHelperTest.sol
@@ -33,17 +33,17 @@ contract TransferHelperTestFakeERC20Compliant {
         shouldRevert = shouldRevert_;
     }
 
-    function transfer(address, uint256) public view returns (bool) {
+    function transfer(address, uint) public view returns (bool) {
         require(!shouldRevert, 'REVERT');
         return success;
     }
 
-    function transferFrom(address, address, uint256) public view returns (bool) {
+    function transferFrom(address, address, uint) public view returns (bool) {
         require(!shouldRevert, 'REVERT');
         return success;
     }
 
-    function approve(address, uint256) public view returns (bool) {
+    function approve(address, uint) public view returns (bool) {
         require(!shouldRevert, 'REVERT');
         return success;
     }
@@ -57,15 +57,15 @@ contract TransferHelperTestFakeERC20Noncompliant {
         shouldRevert = shouldRevert_;
     }
 
-    function transfer(address, uint256) view public {
+    function transfer(address, uint) view public {
         require(!shouldRevert);
     }
 
-    function transferFrom(address, address, uint256) view public {
+    function transferFrom(address, address, uint) view public {
         require(!shouldRevert);
     }
 
-    function approve(address, uint256) view public {
+    function approve(address, uint) view public {
         require(!shouldRevert);
     }
 }
